restructure: tier J — split tests.rs (121 KB) into tests/ subdir by feature
