+++ jetro-core/src/tests/regression.rs
@@ -1,28 +1,12 @@
+//! Mixed-feature regression test corpus: parsing, evaluation, fusion, plan
+//! analysis, fixtures, and assorted backend-comparison checks. The tests in
+//! this file pre-date the feature-specific splits (`chain_write`,
+//! `deep_search`) and are kept together for now.
+
 #[cfg(test)]
 mod tests {
-    use crate::{compile::compiler, parse::parser, vm};
+    use super::super::common::{books, vm_query};
     use serde_json::json;
-
-    fn vm_query(expr: &str, doc: &serde_json::Value) -> Result<serde_json::Value, crate::Error> {
-        let ast = parser::parse(expr)?;
-        let program = compiler::Compiler::compile(&ast, expr);
-        let mut vm = vm::VM::new();
-        Ok(vm.execute(&program, doc)?)
-    }
-
-    fn books() -> serde_json::Value {
-        json!({
-            "store": {
-                "books": [
-                    {"title": "Dune",        "price": 12.99, "rating": 4.8, "genre": "sci-fi",  "tags": ["sci-fi","classic"]},
-                    {"title": "Foundation",  "price":  9.99, "rating": 4.5, "genre": "sci-fi",  "tags": ["sci-fi","series"]},
-                    {"title": "Neuromancer", "price": 11.50, "rating": 4.2, "genre": "cyberpunk","tags": ["sci-fi","cyberpunk"]},
-                    {"title": "1984",        "price":  7.99, "rating": 4.6, "genre": "dystopia", "tags": ["classic","dystopia"]},
-                ]
-            },
-            "user": {"name": "Alice", "age": 30, "score": 85}
-        })
-    }
 
     
     #[test]
@@ -2805,180 +2789,6 @@ mod tests {
         assert_eq!(r, json!({"name": "Bob", "info": {"nested": 1}}));
     }
 
-    
-    #[test]
-    fn patch_simple_field_replace() {
-        let doc = json!({"name": "Alice", "age": 30});
-        let r = vm_query(r#"patch $ { name: "Bob" }"#, &doc).unwrap();
-        assert_eq!(r, json!({"name": "Bob", "age": 30}));
-    }
-
-    #[test]
-    fn patch_nested_field_replace() {
-        let doc = json!({"user": {"name": "Alice", "age": 30}});
-        let r = vm_query(r#"patch $ { user.name: "Bob" }"#, &doc).unwrap();
-        assert_eq!(r, json!({"user": {"name": "Bob", "age": 30}}));
-    }
-
-    #[test]
-    fn patch_delete_field() {
-        let doc = json!({"name": "Alice", "tmp": "remove-me", "age": 30});
-        let r = vm_query(r#"patch $ { tmp: DELETE }"#, &doc).unwrap();
-        assert_eq!(r, json!({"name": "Alice", "age": 30}));
-    }
-
-    #[test]
-    fn patch_add_new_field() {
-        let doc = json!({"name": "Alice"});
-        let r = vm_query(r#"patch $ { age: 42 }"#, &doc).unwrap();
-        assert_eq!(r, json!({"name": "Alice", "age": 42}));
-    }
-
-    #[test]
-    fn patch_wildcard_array() {
-        let doc = json!({"users": [
-            {"name": "Alice", "seen": false},
-            {"name": "Bob",   "seen": false},
-        ]});
-        let r = vm_query(r#"patch $ { users[*].seen: true }"#, &doc).unwrap();
-        assert_eq!(
-            r,
-            json!({"users": [
-                {"name": "Alice", "seen": true},
-                {"name": "Bob",   "seen": true},
-            ]})
-        );
-    }
-
-    #[test]
-    fn patch_wildcard_filter() {
-        let doc = json!({"users": [
-            {"name": "Alice", "active": true,  "role": "user"},
-            {"name": "Bob",   "active": false, "role": "user"},
-            {"name": "Cara",  "active": true,  "role": "user"},
-        ]});
-        let r = vm_query(r#"patch $ { users[* if active].role: "admin" }"#, &doc).unwrap();
-        assert_eq!(
-            r,
-            json!({"users": [
-                {"name": "Alice", "active": true,  "role": "admin"},
-                {"name": "Bob",   "active": false, "role": "user"},
-                {"name": "Cara",  "active": true,  "role": "admin"},
-            ]})
-        );
-    }
-
-    #[test]
-    fn patch_uses_current_value() {
-        let doc = json!({"users": [
-            {"name": "Alice", "email": "ALICE@X"},
-            {"name": "Bob",   "email": "BOB@X"},
-        ]});
-        let r = vm_query(r#"patch $ { users[*].email: @.lower() }"#, &doc).unwrap();
-        assert_eq!(
-            r,
-            json!({"users": [
-                {"name": "Alice", "email": "alice@x"},
-                {"name": "Bob",   "email": "bob@x"},
-            ]})
-        );
-    }
-
-    #[test]
-    fn patch_conditional_when_truthy() {
-        let doc = json!({"count": 5, "enabled": true});
-        let r = vm_query(r#"patch $ { count: @ + 1 when $.enabled }"#, &doc).unwrap();
-        assert_eq!(r, json!({"count": 6, "enabled": true}));
-    }
-
-    #[test]
-    fn patch_conditional_when_falsy_skips() {
-        let doc = json!({"count": 5, "enabled": false});
-        let r = vm_query(r#"patch $ { count: @ + 1 when $.enabled }"#, &doc).unwrap();
-        assert_eq!(r, json!({"count": 5, "enabled": false}));
-    }
-
-    #[test]
-    fn patch_multiple_ops_in_order() {
-        let doc = json!({"a": 1, "b": 2, "c": 3});
-        let r = vm_query(r#"patch $ { a: 10, b: DELETE, c: 30 }"#, &doc).unwrap();
-        assert_eq!(r, json!({"a": 10, "c": 30}));
-    }
-
-    #[test]
-    fn patch_index_access() {
-        let doc = json!({"items": [10, 20, 30]});
-        let r = vm_query(r#"patch $ { items[1]: 99 }"#, &doc).unwrap();
-        assert_eq!(r, json!({"items": [10, 99, 30]}));
-    }
-
-    #[test]
-    fn patch_delete_from_wildcard() {
-        let doc = json!({"users": [
-            {"name": "Alice", "active": true},
-            {"name": "Bob",   "active": false},
-            {"name": "Cara",  "active": true},
-        ]});
-        let r = vm_query(r#"patch $ { users[* if not active]: DELETE }"#, &doc).unwrap();
-        assert_eq!(
-            r,
-            json!({"users": [
-                {"name": "Alice", "active": true},
-                {"name": "Cara",  "active": true},
-            ]})
-        );
-    }
-
-    #[test]
-    fn patch_composes_pipe() {
-        let doc = json!({"name": "Alice", "age": 30});
-        let r = vm_query(r#"patch $ { name: "Bob" } | @.name"#, &doc).unwrap();
-        assert_eq!(r, json!("Bob"));
-    }
-
-    #[test]
-    fn patch_composes_method_chain() {
-        let doc = json!({"name": "Alice", "age": 30});
-        let r = vm_query(r#"patch $ { name: "Bob" }.keys()"#, &doc).unwrap();
-        let mut keys = r.as_array().unwrap().clone();
-        keys.sort_by(|a, b| a.as_str().unwrap().cmp(b.as_str().unwrap()));
-        assert_eq!(keys, vec![json!("age"), json!("name")]);
-    }
-
-    #[test]
-    fn patch_composes_nested_in_object() {
-        let doc = json!({"name": "Alice"});
-        let r = vm_query(r#"{result: patch $ { name: "Bob" }}"#, &doc).unwrap();
-        assert_eq!(r, json!({"result": {"name": "Bob"}}));
-    }
-
-    #[test]
-    fn patch_composes_let_binding() {
-        let doc = json!({"name": "Alice", "age": 30});
-        let r = vm_query(r#"let x = patch $ { name: "Bob" } in x.name"#, &doc).unwrap();
-        assert_eq!(r, json!("Bob"));
-    }
-
-    #[test]
-    fn patch_composes_nested_patch() {
-        let doc = json!({"name": "Alice", "age": 30});
-        let r = vm_query(r#"patch (patch $ { name: "Bob" }) { age: 99 }"#, &doc).unwrap();
-        assert_eq!(r, json!({"name": "Bob", "age": 99}));
-    }
-
-    #[test]
-    fn patch_composes_inside_map() {
-        let doc = json!({"users": [{"n": 1}, {"n": 2}, {"n": 3}]});
-        let r = vm_query(r#"$.users.map(patch @ { n: @ * 10 })"#, &doc).unwrap();
-        assert_eq!(r, json!([{"n": 10}, {"n": 20}, {"n": 30}]));
-    }
-
-    #[test]
-    fn patch_delete_mark_outside_patch_errors() {
-        let doc = json!({});
-        let r = vm_query(r#"DELETE"#, &doc);
-        assert!(r.is_err());
-    }
 
     
     fn saas() -> serde_json::Value {
@@ -3277,351 +3087,6 @@ mod tests {
         assert!(titles.contains(&json!("Dune")));
     }
 
-    #[test]
-    fn simd_scan_descendant_matches_tree_walker() {
-        use crate::Jetro;
-        let raw = br#"{"a":{"test":1},"b":[{"test":2},{"other":9},{"test":3}],"comment":"the \"test\": lie"}"#.to_vec();
-        let j_bytes = Jetro::from_bytes(raw.clone()).unwrap();
-        let j_tree = Jetro::new(serde_json::from_slice(&raw).unwrap());
-        assert_eq!(
-            j_bytes.collect("$..test").unwrap(),
-            j_tree.collect("$..test").unwrap()
-        );
-    }
-
-    #[test]
-    fn simd_scan_chains_further_steps() {
-        use crate::Jetro;
-        let raw =
-            br#"{"users":[{"id":1,"name":"a"},{"id":2,"name":"b"},{"id":3,"name":"c"}]}"#.to_vec();
-        let j = Jetro::from_bytes(raw).unwrap();
-        let r = j.collect("$..id.sum()").unwrap();
-        assert_eq!(r, json!(6));
-    }
-
-    #[test]
-    fn simd_scan_via_vm_path() {
-        
-        
-        use crate::Jetro;
-        let raw = br#"{"a":{"x":1},"b":[{"x":2},{"x":3}]}"#.to_vec();
-        let j_b = Jetro::from_bytes(raw.clone()).unwrap();
-        let j_t = Jetro::new(serde_json::from_slice(&raw).unwrap());
-        
-        assert_eq!(j_b.collect("$..x").unwrap(), j_t.collect("$..x").unwrap());
-        assert_eq!(j_b.collect("$..x").unwrap(), j_t.collect("$..x").unwrap());
-    }
-
-    #[test]
-    fn simd_scan_vm_path_aggregate() {
-        
-        
-        use crate::Jetro;
-        let raw = br#"{"rows":[{"p":10},{"p":20},{"p":30}]}"#.to_vec();
-        let j = Jetro::from_bytes(raw).unwrap();
-        assert_eq!(j.collect("$..p.sum()").unwrap(), json!(60));
-    }
-
-    #[test]
-    fn simd_scan_literal_eq_int() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({"xs":[{"n":10},{"n":42},{"n":10},{"n":42},{"n":7}]});
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j = Jetro::from_bytes(raw).unwrap();
-        assert_eq!(j.collect("$..n.filter(@ == 42)").unwrap(), json!([42, 42]));
-    }
-
-    #[test]
-    fn simd_scan_literal_eq_string() {
-        use crate::Jetro;
-        let doc = json!({
-            "events":[
-                {"type":"action"},{"type":"idle"},
-                {"type":"action"},{"type":"noop"}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j = Jetro::from_bytes(raw).unwrap();
-        assert_eq!(
-            j.collect(r#"$..type.filter(@ == "action")"#).unwrap(),
-            json!(["action", "action"])
-        );
-    }
-
-    #[test]
-    fn simd_scan_literal_eq_bool_null() {
-        use crate::Jetro;
-        let doc = json!({"xs":[{"v":true},{"v":false},{"v":true},{"v":null}]});
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j = Jetro::from_bytes(raw.clone()).unwrap();
-        assert_eq!(
-            j.collect("$..v.filter(@ == true)").unwrap(),
-            json!([true, true])
-        );
-
-        let j2 = Jetro::from_bytes(raw).unwrap();
-        assert_eq!(j2.collect("$..v.filter(@ == null)").unwrap(), json!([null]));
-    }
-
-    #[test]
-    fn route_c_chained_descendants_match_tree_walker() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({
-            "outer":[
-                {"inner":[{"leaf":1},{"leaf":2}]},
-                {"inner":[{"leaf":3},{"leaf":4}]}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = "$..outer.first()..inner.first()..leaf";
-        assert_eq!(j_b.collect(q).unwrap(), j_t.collect(q).unwrap());
-        assert_eq!(j_b.collect(q).unwrap(), json!([1, 2]));
-    }
-
-    #[test]
-    fn route_c_descendant_after_filter_eq() {
-        
-        use crate::Jetro;
-        let doc = json!({
-            "items":[
-                {"kind":"a","children":[{"v":1},{"v":2}]},
-                {"kind":"b","children":[{"v":3},{"v":4}]}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..kind.filter(@ == "a")"#;
-        assert_eq!(j_b.collect(q).unwrap(), j_t.collect(q).unwrap());
-        assert_eq!(j_b.collect(q).unwrap(), json!(["a"]));
-    }
-
-    #[test]
-    fn route_c_quantifier_scalar_result() {
-        
-        use crate::Jetro;
-        let doc = json!({"xs":[{"id":7},{"id":8}]});
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        assert_eq!(j_b.collect("$..id.first()").unwrap(), json!(7));
-    }
-
-    #[test]
-    fn deep_find_field_eq_scan_matches_tree_walker() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({
-            "a":[
-                {"type":"action","v":1},
-                {"type":"idle","v":2},
-                {"nested":{"type":"action","v":3}}
-            ],
-            "b":{"type":"action","v":4}
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..find(@.type == "action")"#;
-        let rb = j_b.collect(q).unwrap();
-        let rt = j_t.collect(q).unwrap();
-        assert_eq!(rb, rt);
-        let arr = rb.as_array().unwrap();
-        assert_eq!(arr.len(), 3);
-    }
-
-    #[test]
-    fn deep_find_field_eq_int_literal() {
-        use crate::Jetro;
-        let doc = json!({"xs":[{"id":1,"t":"a"},{"id":2,"t":"b"},{"deep":{"id":1,"t":"c"}}]});
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = "$..find(@.id == 1)";
-        assert_eq!(j_b.collect(q).unwrap(), j_t.collect(q).unwrap());
-        assert_eq!(j_b.collect(q).unwrap().as_array().unwrap().len(), 2);
-    }
-
-    #[test]
-    fn deep_find_field_eq_chains_further() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({
-            "items":[
-                {"type":"action","v":10},
-                {"type":"idle","v":20},
-                {"type":"action","v":30}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..find(@.type == "action").map(v)"#;
-        assert_eq!(j_b.collect(q).unwrap(), j_t.collect(q).unwrap());
-        assert_eq!(j_b.collect(q).unwrap(), json!([10, 30]));
-    }
-
-    #[test]
-    fn find_shallow_multi_pred_and() {
-        
-        use crate::Jetro;
-        let doc = json!({"xs":[
-            {"t":"a","v":1},
-            {"t":"a","v":2},
-            {"t":"b","v":1}
-        ]});
-        let j = Jetro::new(doc);
-        let r = j.collect(r#"$.xs.find(@.t == "a", @.v == 1)"#).unwrap();
-        assert_eq!(r, json!([{"t":"a","v":1}]));
-    }
-
-    #[test]
-    fn find_shallow_single_pred_still_works() {
-        use crate::Jetro;
-        let doc = json!({"xs":[{"v":1},{"v":2}]});
-        let j = Jetro::new(doc);
-        let r = j.collect(r#"$.xs.find(@.v == 2)"#).unwrap();
-        assert_eq!(r, json!([{"v":2}]));
-    }
-
-    #[test]
-    fn deep_find_multi_pred_matches_scan_and_tree() {
-        
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","device":"mobile","id":1},
-                {"type":"action","device":"desktop","id":2},
-                {"type":"idle","device":"mobile","id":3},
-                {"nested":{"type":"action","device":"mobile","id":4}}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..find(@.type == "action", @.device == "mobile")"#;
-        let rb = j_b.collect(q).unwrap();
-        let rt = j_t.collect(q).unwrap();
-        assert_eq!(rb, rt);
-        assert_eq!(rb.as_array().unwrap().len(), 2);
-    }
-
-    #[test]
-    fn deep_find_then_filter_eq_refines_spans() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","device":"mobile","state":"on","id":1},
-                {"type":"action","device":"mobile","state":"off","id":2},
-                {"type":"action","device":"desktop","state":"on","id":3}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..find(@.type == "action", @.device == "mobile").filter(@.state == "on")"#;
-        let rb = j_b.collect(q).unwrap();
-        let rt = j_t.collect(q).unwrap();
-        assert_eq!(rb, rt);
-        assert_eq!(rb.as_array().unwrap().len(), 1);
-    }
-
-    #[test]
-    fn deep_find_then_filter_cmp_refines_spans() {
-        
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","v":1},
-                {"type":"action","v":10},
-                {"type":"action","v":100},
-                {"type":"idle","v":200}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let j_t = Jetro::new(doc);
-        let q = r#"$..find(@.type == "action").filter(@.v > 5).map(v).sum()"#;
-        let rb = j_b.collect(q).unwrap();
-        let rt = j_t.collect(q).unwrap();
-        assert_eq!(rb, rt);
-        assert_eq!(rb, json!(110));
-    }
-
-    #[test]
-    fn deep_find_then_filter_then_count() {
-        
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","v":1},
-                {"type":"action","v":10},
-                {"type":"action","v":100},
-                {"type":"idle","v":50}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let q = r#"$..find(@.type == "action").filter(@.v >= 10).count()"#;
-        assert_eq!(j_b.collect(q).unwrap(), json!(2));
-    }
-
-    #[test]
-    fn deep_find_then_fused_filter_map_sum() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","v":1},
-                {"type":"action","v":10},
-                {"type":"action","v":100}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        let q = r#"$..find(@.type == "action").filter(@.v > 5).map(v).sum()"#;
-        assert_eq!(j_b.collect(q).unwrap(), json!(110));
-    }
-
-    #[test]
-    fn deep_find_then_fused_filter_map_array() {
-        use crate::Jetro;
-        let doc = json!({
-            "rows":[
-                {"type":"action","v":1},
-                {"type":"action","v":10},
-                {"type":"idle","v":50}
-            ]
-        });
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw.clone()).unwrap();
-        let j_t = Jetro::new(serde_json::from_slice::<serde_json::Value>(&raw).unwrap());
-        let q = r#"$..find(@.type == "action").filter(@.v >= 10).map(v)"#;
-        assert_eq!(j_b.collect(q).unwrap(), j_t.collect(q).unwrap());
-        assert_eq!(j_b.collect(q).unwrap(), json!([10]));
-    }
-
-    
-    #[test]
-    fn route_c_one_mismatch_errors_via_fallthrough() {
-        
-        
-        use crate::Jetro;
-        let doc = json!({"xs":[{"id":1},{"id":2}]});
-        let raw = serde_json::to_vec(&doc).unwrap();
-        let j_b = Jetro::from_bytes(raw).unwrap();
-        assert!(j_b.collect("$..id.one()").is_err());
-    }
 
     
     #[test]
